Refactor LfContingency
